+++ server.js
@@ -12,24 +12,7 @@ const PORT = process.env.PORT || 3001;
 
 const hbs = exphbs.create();
 
-const sess = {
-  secret: 'Super secret secret',
-
-cookie: {
-  maxAge: 24 * 60 * 60 * 1000,
-  httpOnly: true,
-  secure: false,
-  sameSite: 'strict',
-},
-
-  resave: false,
-  saveUninitialized: true,
-  store: new SequelizeStore({
-    db: sequelize
-  })
-};
-
-app.use(session(sess));
+app.use(sessionMiddleware);
 
 app.engine('handlebars', hbs.engine);
 app.set('view engine', 'handlebars');
